+++ refgenconf/helpers.py
@@ -1,11 +1,13 @@
 """ Helper functions """
 
 import os
-import yacman
-from .const import CFG_ENV_VARS
+from yacman import select_config
+from .const import CFG_ENV_VARS, BUILD_STATS_DIR
+from re import sub
+from ubiquerg import is_command_callable
 
 
-__all__ = ["select_genome_config"]
+__all__ = ["select_genome_config", "get_dir_digest"]
 
 
 def select_genome_config(filename=None, conf_env_vars=CFG_ENV_VARS, **kwargs):
@@ -17,7 +19,7 @@ def select_genome_config(filename=None, conf_env_vars=CFG_ENV_VARS, **kwargs):
         consider; basically, a prioritized search list
     :return str: path to genome configuration file
     """
-    return yacman.select_config(filename, conf_env_vars, **kwargs)
+    return select_config(filename, conf_env_vars, **kwargs)
 
 
 def unbound_env_vars(path):
@@ -38,3 +40,32 @@ def unbound_env_vars(path):
 def asciify_json_dict(json_dict):
     from ubiquerg.collection import asciify_dict
     return asciify_dict(json_dict)
+
+
+def get_dir_digest(path, pm=None):
+    """
+    Generate a MD5 digest that reflects just the contents of the
+    files in the selected directory.
+
+    :param str path: path to the directory to digest
+    :param pypiper.PipelineManager pm: a pipeline object, optional.
+    The subprocess module will be used if not provided
+    :return str: a digest, e.g. a3c46f201a3ce7831d85cf4a125aa334
+    """
+    if not is_command_callable("md5sum"):
+        raise OSError("md5sum command line tool is required for asset digest "
+                      "calculation. \n"
+                      "Install and try again, e.g on macOS: 'brew install "
+                      "md5sha1sum'")
+    cmd = "cd {}; find . -type f -not -path './" + BUILD_STATS_DIR + \
+          "*' -exec md5sum {{}} \; | sort -k 2 | awk '{{print $1}}' | md5sum"
+    try:
+        x = pm.checkprint(cmd.format(path))
+    except AttributeError:
+        try:
+            from subprocess import check_output
+            x = check_output(cmd.format(path), shell=True).decode("utf-8")
+        except Exception as e:
+
+            return
+    return str(sub(r'\W+', '', x))  # strips non-alphanumeric